+++ client/src/features/results/ConceptSnippet.tsx
@@ -50,6 +50,7 @@ const ConceptSnippet: React.FC<IProps> = ({
     const concept = new ConceptParser(data)
     const images = concept.getImages()
     const descriptions = concept.getDescriptions(config.aat.langen)
+    const className = concept.getEntityClassName()
 
     const snippetDataComponent = (
       <React.Fragment>
@@ -62,6 +63,16 @@ const ConceptSnippet: React.FC<IProps> = ({
                   0,
                   view === 'list' ? 200 : 100,
                 )}...`}</StyledDd>
+              </Col>
+            </Row>
+          )}
+          {className && (
+            <Row>
+              <Col>
+                <StyledDt>Concept Class</StyledDt>
+                <StyledDd data-testid="concept-snippet-class">
+                  {className}
+                </StyledDd>
               </Col>
             </Row>
           )}